@@ -2,107 +2,121 @@ apply plugin: 'java'
 apply plugin: 'idea'
 apply plugin: 'eclipse'
 apply plugin: 'groovy'
-apply plugin: 'maven'
+apply plugin: 'maven-publish'
 apply plugin: 'signing'
 
 apply plugin: org.owasp.dependencycheck.gradle.DependencyCheckPlugin
 
 group = 'io.elastic'
 version = '3.3.10-SNAPSHOT'
-ext.isReleaseVersion = !version.endsWith("SNAPSHOT")
 
-sourceCompatibility = 1.8
-targetCompatibility = 1.8
-
-signing {
-    required { isReleaseVersion }
-    sign configurations.archives
-}
+sourceCompatibility = 18
+targetCompatibility = 18
 
 repositories {
     mavenCentral()
-    maven {
-        url "https://oss.sonatype.org/content/repositories/snapshots"
-    }
 }
 
 dependencies {
-    compile "org.glassfish:javax.json:1.0.4"
-    compile "com.rabbitmq:amqp-client:5.9.0"
-    compile "commons-codec:commons-codec:1.10"
-    compile "commons-io:commons-io:2.7"
-    compile "org.apache.httpcomponents:httpclient:4.5.13"
-    compile "com.google.inject:guice:5.1.0"
-    compile "com.google.inject.extensions:guice-assistedinject:5.1.0"
-    compile "ch.qos.logback:logback-classic:1.2.10"
-    compile "ch.qos.logback.contrib:logback-json-classic:0.1.5"
-    compile "ch.qos.logback.contrib:logback-jackson:0.1.5"
-    compile "com.fasterxml.jackson.core:jackson-databind:2.13.2.2"
+    implementation "org.glassfish:javax.json:1.0.4"
+    implementation "com.rabbitmq:amqp-client:5.9.0"
+    implementation "commons-codec:commons-codec:1.10"
+    implementation "commons-io:commons-io:2.7"
+    implementation "org.apache.httpcomponents:httpclient:4.5.13"
+    implementation "com.google.inject:guice:5.1.0"
+    implementation "com.google.inject.extensions:guice-assistedinject:5.1.0"
+    implementation "ch.qos.logback:logback-classic:1.2.10"
+    implementation "ch.qos.logback.contrib:logback-json-classic:0.1.5"
+    implementation "ch.qos.logback.contrib:logback-jackson:0.1.5"
+    implementation "com.fasterxml.jackson.core:jackson-databind:2.13.2.2"
 
-    testRuntime 'org.codehaus.groovy:groovy-all:2.0.5'
-    testCompile "org.spockframework:spock-core:1.3-groovy-2.4"
-    testCompile "com.github.rest-driver:rest-client-driver:2.0.0"
-    testCompile "org.eclipse.jetty:jetty-server:9.4.1.v20170120"
-    testCompile 'com.github.tomakehurst:wiremock-standalone:2.20.0'
+    testRuntimeOnly 'org.codehaus.groovy:groovy-all:2.4.15'
+    testImplementation "org.spockframework:spock-core:1.3-groovy-2.4"
+    testImplementation "com.github.rest-driver:rest-client-driver:2.0.0"
+    testImplementation "org.eclipse.jetty:jetty-server:9.4.1.v20170120"
+    testImplementation 'com.github.tomakehurst:wiremock-jre8-standalone:2.33.2'
 }
 
-uploadArchives {
-    repositories {
-        mavenLocal()
-        if (project.hasProperty('sonatypeUsername') && project.hasProperty('sonatypePassword')) {
-            mavenDeployer {
-                beforeDeployment { MavenDeployment deployment -> signing.signPom(deployment) }
+def isReleaseBuild() {
+    return !version.endsWith("SNAPSHOT")
+}
 
-                repository(url: "https://oss.sonatype.org/service/local/staging/deploy/maven2/") {
-                    authentication(userName: sonatypeUsername, password: sonatypePassword)
+def getReleaseRepositoryUrl() {
+    return "https://oss.sonatype.org/service/local/staging/deploy/maven2/"
+}
+
+def getSnapshotRepositoryUrl() {
+    return "https://oss.sonatype.org/content/repositories/snapshots/"
+}
+
+def getRepositoryUsername() {
+    return project.hasProperty('sonatypeUsername') ? sonatypeUsername : ""
+}
+
+def getRepositoryPassword() {
+    return project.hasProperty('sonatypePassword') ? sonatypePassword : ""
+}
+
+publishing {
+    publications {
+        mavenJava(MavenPublication) {
+            pom {
+                name = 'elasticio-sailor-jvm'
+                packaging = 'jar'
+                description = 'Sailor for JVM'
+                url = 'http://www.elastic.io'
+                inceptionYear = '2015'
+
+                organization {
+                    name = 'elastic.io GmbH'
+                    url = 'http://www.elastic.io'
                 }
 
-                snapshotRepository(url: "https://oss.sonatype.org/content/repositories/snapshots/") {
-                    authentication(userName: sonatypeUsername, password: sonatypePassword)
+                scm {
+                    url = 'scm:git@github.com:elasticio/sailor-jvm.git'
+                    connection = 'scm:git@github.com:elasticio/sailor-jvm.git'
+                    developerConnection = 'scm:git@github.com:elasticio/sailor-jvm.git'
                 }
 
-                pom.project {
-                    name 'elasticio-sailor-jvm'
-                    packaging 'jar'
-                    description 'Sailor for JVM'
-                    url 'http://www.elastic.io'
-                    inceptionYear '2015'
-
-                    organization {
-                        name 'elastic.io GmbH'
-                        url 'http://www.elastic.io'
+                licenses {
+                    license {
+                        name = 'The Apache Software License, Version 2.0'
+                        url = 'http://www.apache.org/licenses/LICENSE-2.0.txt'
+                        distribution = 'repo'
                     }
+                }
 
-                    scm {
-                        url 'scm:git@github.com:elasticio/sailor-jvm.git'
-                        connection 'scm:git@github.com:elasticio/sailor-jvm.git'
-                        developerConnection 'scm:git@github.com:elasticio/sailor-jvm.git'
-                    }
+                issueManagement {
+                    system = 'GitHub'
+                    url = 'https://github.com/elasticio/sailor-jvm/issues'
+                }
 
-                    licenses {
-                        license {
-                            name 'The Apache Software License, Version 2.0'
-                            url 'http://www.apache.org/licenses/LICENSE-2.0.txt'
-                            distribution 'repo'
-                        }
-                    }
-
-                    issueManagement {
-                        system 'GitHub'
-                        url 'https://github.com/elasticio/sailor-jvm/issues'
-                    }
-
-                    developers {
-                        developer {
-                            id 'elasticio'
-                            name 'elastic.io GmbH'
-                            email 'support@elastic.io'
-                        }
+                developers {
+                    developer {
+                        id = 'elasticio'
+                        name = 'elastic.io GmbH'
+                        email = 'support@elastic.io'
                     }
                 }
             }
         }
     }
+
+    repositories {
+        maven {
+            url = isReleaseBuild() ? getReleaseRepositoryUrl() : getSnapshotRepositoryUrl()
+
+            credentials {
+                username = getRepositoryUsername()
+                password = getRepositoryPassword()
+            }
+        }
+    }
+}
+
+signing {
+    required { isReleaseBuild() && gradle.taskGraph.hasTask("publish") }
+    sign publishing.publications.mavenJava
 }
 
 task sourceJar(type: Jar) {
